[portal] Add "jobs.py compare" (#63)
This command makes it a little easier to compare two or more results files.  You can give it a request ID, a request directory, or a filename.  It also supports comparing with results uploaded to [the ideas repo](https://github.com/faster-cpython/ideas/tree/main/benchmark-results).  Give it the relative path of one of the uploaded filenames, the upload ID (the basename without suffix), or even a version number.

This also makes a number of small fixes, as well as reporting "main" as "3.12.0a0".

* Move PyperfResults to its own module.

* Move FasterCPythonResults to _pyperformance.

* Add _utils.resolve_os_name() and _utils.resolve_cpu_arch().

* Add PythonImplementation.

* Support a "positive" range in _utils.validte_int().

* Fix some uses of __new__().

* improvements to _utils.Version.

* Move the compile-bench job impl to its own module.

* Add PyperfUploadID, PyperfResultsFile, and PyperfResultsMetadata.

* Merge in per-benchmark metadata.

* Fix the upload version for main.

* Add the "compare" command.

* Support comparing uploaded results.