Support the Mac worker machine
This adds a new arg `--worker` to select the worker machine.
It looks up a configuration file from `BENCH/WORKERS/{worker_name}` and
then uses a worker-specific queue in `BENCH/QUEUES/{worker_name}`.

The rest of the changes are related to keeping paths in "tilde-including" form
as long as possible for any paths that might be sent to the worker machine,
since tilde-expansion is different on Linux and Mac.